型の不一致による警告を修正

+++ KxSMBSample/TreeViewController.m
@@ -258,7 +258,7 @@
 
 - (void) actionCopyFile:(id)sender
 {
-    NSString *name = [NSString stringWithFormat:@"%d.tmp", (NSUInteger)[NSDate timeIntervalSinceReferenceDate]];
+    NSString *name = [NSString stringWithFormat:@"%lu.tmp", (unsigned long)[NSDate timeIntervalSinceReferenceDate]];
     NSString *path = [_path stringByAppendingSMBPathComponent:name];
     
     KxSMBProvider *provider = [KxSMBProvider sharedSmbProvider];
